chore: rework helper functions and operations
