+++ components/script/dom/servoparser/mod.rs
@@ -507,10 +507,6 @@ impl ServoParser {
         self.tokenizer.get_current_line()
     }
 
-    pub(crate) fn get_document(&self) -> DomRoot<Document> {
-        self.document.as_rooted()
-    }
-
     #[cfg_attr(crown, expect(crown::unrooted_must_root))]
     fn new_inherited(
         document: &Document,
@@ -967,7 +963,7 @@ impl ParserContext {
     pub(crate) fn get_document(&self) -> Option<DomRoot<Document>> {
         self.parser
             .as_ref()
-            .map(|parser| parser.root().get_document())
+            .map(|parser| parser.root().document.as_rooted())
     }
 
     /// <https://html.spec.whatwg.org/multipage/#creating-a-policy-container-from-a-fetch-response>
